# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreamFlix API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List all movies and series
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: genre
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [movie, series]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentListResponse"
    post:
      operationId: createContent
      tags: [content]
      summary: Create movie or series (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
...
  /content/latest:
    get:
      operationId: getLatestContent
      tags: [content]
      summary: Get latest additions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Latest content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Content"
  /content/popular:
    get:
      operationId: getPopularContent
      tags: [content]
      summary: Get popular content by views and ratings
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Popular content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Content"
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get movie or series by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update movie or series (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SEASONS
  /content/{contentId}/seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List seasons for a series
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create a season (admin)
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{seasonId}:
    patch:
      operationId: updateSeason
      tags: [seasons]
      summary: Update season (admin)
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
    delete:
      operationId: deleteSeason
      tags: [seasons]
      summary: Delete season (admin)
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # EPISODES
  /seasons/{seasonId}/episodes:
    get:
      operationId: listEpisodes
      tags: [episodes]
      summary: List episodes for a season
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of episodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"
    post:
      operationId: createEpisode
      tags: [episodes]
      summary: Create an episode (admin)
      parameters:
        - name: seasonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
  /episodes/{episodeId}:
    patch:
      operationId: updateEpisode
      tags: [episodes]
      summary: Update episode (admin)
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
    delete:
      operationId: deleteEpisode
      tags: [episodes]
      summary: Delete episode (admin)
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FAVORITES
...
components:
  schemas:
...
    Content:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        backdropUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        durationMinutes:
          type: ["integer", "null"]
        contentType:
          type: string
          enum: [movie, series]
        isFeatured:
          type: boolean
        averageRating:
          type: ["number", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - contentType
        - isFeatured
        - viewCount
        - createdAt
    ContentDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        backdropUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        releaseYear:
          type: ["integer", "null"]
        durationMinutes:
          type: ["integer", "null"]
        contentType:
          type: string
          enum: [movie, series]
        isFeatured:
          type: boolean
        averageRating:
          type: ["number", "null"]
        viewCount:
          type: integer
        createdAt:
          type: string
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/SeasonWithEpisodes"
        isFavorite:
          type: boolean
        userRating:
          type: ["integer", "null"]
        watchProgress:
          type: ["integer", "null"]
      required:
        - id
        - title
        - contentType
        - isFeatured
        - viewCount
        - createdAt
        - seasons
        - isFavorite
    ContentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Content"
        total:
          type: integer
      required:
        - items
        - total
    ContentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        posterUrl:
          type: string
        backdropUrl:
          type: string
        videoUrl:
          type: string
        categoryId:
          type: integer
        genre:
          type: string
        releaseYear:
          type: integer
        durationMinutes:
          type: integer
        contentType:
          type: string
          enum: [movie, series]
        isFeatured:
          type: boolean
      required:
        - title
        - contentType
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        posterUrl:
          type: string
        backdropUrl:
          type: string
        videoUrl:
          type: string
        categoryId:
          type: integer
        genre:
          type: string
        releaseYear:
          type: integer
        durationMinutes:
          type: integer
        contentType:
          type: string
          enum: [movie, series]
        isFeatured:
          type: boolean
    Season:
      type: object
      properties:
        id:
          type: integer
        contentId:
          type: integer
        seasonNumber:
          type: integer
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - contentId
        - seasonNumber
        - createdAt
    SeasonWithEpisodes:
      type: object
      properties:
        id:
          type: integer
        contentId:
          type: integer
        seasonNumber:
          type: integer
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
      required:
        - id
        - contentId
        - seasonNumber
        - createdAt
        - episodes
    SeasonInput:
      type: object
      properties:
        seasonNumber:
          type: integer
        title:
          type: string
        description:
          type: string
      required:
        - seasonNumber
    SeasonUpdate:
      type: object
      properties:
        seasonNumber:
          type: integer
        title:
          type: string
        description:
          type: string
    Episode:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        episodeNumber:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - episodeNumber
        - title
        - createdAt
    EpisodeInput:
      type: object
      properties:
        episodeNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        durationMinutes:
          type: integer
      required:
        - episodeNumber
        - title
    EpisodeUpdate:
      type: object
      properties:
        episodeNumber:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        durationMinutes:
          type: integer
...