# lib/api-zod/src/generated/types/seasonUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StreamFlix API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SeasonUpdate {
  seasonNumber?: number;
  title?: string;
  description?: string;
}